+++ state-machines/binary/src/binary_basic_collector.rs
@@ -72,7 +72,7 @@ impl BusDevice<u64> for BinaryBasicCollector {
         _pending: &mut VecDeque<(BusId, Vec<u64>)>,
     ) -> bool {
         debug_assert!(*bus_id == OPERATION_BUS_ID);
-        let instance_complete = self.inputs.len() == self.num_operations as usize;
+        let instance_complete = self.inputs.len() == self.num_operations;
 
         if instance_complete && !self.force_execute_to_end {
             return false;
@@ -108,7 +108,7 @@ impl BusDevice<u64> for BinaryBasicCollector {
         }
         self.inputs.push(BinaryInput::from(&op_data));
 
-        self.inputs.len() < self.num_operations as usize || self.force_execute_to_end
+        self.inputs.len() < self.num_operations || self.force_execute_to_end
     }
 
     /// Returns the bus IDs associated with this instance.
